Move `normaliseString()` + tests to separate files
This prevents circular dependency issues
